requirements.txt
Flask==0.12.2
Flask-Session==0.3.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
SQLAlchemy==1.1.11
Werkzeug==0.12.2
